fix: stat record creation blocking new post creation
Creating a new stat record was returning a 500 error when creating a new
post. This was due to an incorrect parameter in the payload.

The stat creation error has been refactored into a different try/catch
block with error reporting. The reason is that it should not block the
creation of a new post anyway, as a record will be created when needed,
later.