refactor: break admin menu into multiple components
Since we now have multiple contexts for posts (i.e. a post by id or a
post by index in a tag), the update/delete menu no longer works.

The only place to correctly display such a menu is now with the post
itself, where it is placed in the page.

For consistency, the equivalent Page menu has also been moved.